+++ src/components/StringList/constants.ts
